requirements.txt
websocket-client
astroid			
Click			
colorama		
fastapi			
gunicorn		
h11				
isort			
lazy-object-proxy
mccabe			
pydantic		
pylint			
six				
starlette		
uvicorn			
websockets		
wrapt			
uvloop
fastapi
requests
playwright
celery
python-dotenv
kombu
redis